pay from wallet option added

+++ lib/models/payment/pay_to_user_request.dart
@@ -2,7 +2,7 @@ class PayToUserRequest {
   int? customerProfileId;
   int? paymentProfile;
   double? amount;
-  int? walletAmount;
+  double? walletAmount;
   int? toUserId;
   String? paymentMethod;
   int? paymentReqId;
